[IMP] website_pricelist_medical_insurance: Rename website_medical_insurance_pricelist * Rename module to correspond with OCA naming best practices
