--- website_medical_insurance_pricelist/tests/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import test_website
