+++ website_pricelist_medical_insurance/security/ir.model.access.csv
@@ -1,0 +1,5 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_medical_insurance_template_portal,medical.insurance.template.portal,medical_insurance.model_medical_insurance_template,base.group_portal,1,0,0,0
+access_medical_insurance_template_public,medical.insurance.template.public,medical_insurance.model_medical_insurance_template,base.group_public,1,0,0,0
+access_medical_insurance_template_plan_portal,medical.insurance.plan.portal,medical_insurance.model_medical_insurance_plan,base.group_portal,1,0,0,0
+access_medical_insurance_template_plan_public,medical.insurance.plan.public,medical_insurance.model_medical_insurance_plan,base.group_public,1,0,0,0
